D4CRIS-416 first implementation of the contract (missing bitstream move operation and replace operation on attribute at accessConditions level)

+++ dspace-spring-rest/src/main/java/org/dspace/app/rest/submit/factory/impl/LicenseAddPatchOperation.java
@@ -7,6 +7,7 @@
  */
 package org.dspace.app.rest.submit.factory.impl;
 
+import org.apache.commons.lang.BooleanUtils;
 import org.dspace.content.Item;
 import org.dspace.content.LicenseUtils;
 import org.dspace.content.WorkspaceItem;
@@ -19,15 +20,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 /**
  * Submission "add" PATCH operation
  * 
- * To accept a license the timestamp of the acceptance the client must send a
- * JSON Patch ADD operation as follow.
+ * To accept/reject the license.
  * 
  * Example: <code>
  * curl -X PATCH http://${dspace.url}/dspace-spring-rest/api/submission/workspaceitems/31599 -H "Content-Type: application/json" -d '[{ "op": "add", "path": "/sections/license/granted", "value":"true"}]'
  * </code>
  * 
  * Please note that according to the JSON Patch specification RFC6902 a
- * subsequent add operation on the acceptanceDate will have the effect to
+ * subsequent add operation on the "granted" path will have the effect to
  * replace the previous granted license with a new one.
  * 
  * 
@@ -52,17 +52,26 @@ public class LicenseAddPatchOperation extends AddPatchOperation<String> {
 	@Override
 	void add(Context context, Request currentRequest, WorkspaceItem source, String path, Object value)
 			throws Exception {
+		
+		Boolean grant = BooleanUtils.toBooleanObject((String)value);
+		
+		if(grant==null) {
+			throw new IllegalArgumentException("Value is not a valid boolean expression (permitted value: on/off, true/false and yes/no");
+		}
+		
 		Item item = source.getItem();
 		EPerson submitter = context.getCurrentUser();
 
 		// remove any existing DSpace license (just in case the user
 		// accepted it previously)
 		itemService.removeDSpaceLicense(context, item);
-
-		String license = LicenseUtils.getLicenseText(context.getCurrentLocale(), source.getCollection(), item,
+		
+		if(grant) {
+			String license = LicenseUtils.getLicenseText(context.getCurrentLocale(), source.getCollection(), item,
 				submitter);
 
-		LicenseUtils.grantLicense(context, item, license, null);
+			LicenseUtils.grantLicense(context, item, license, null);
+		}
 	}
 
 }
