+++ dspace-spring-rest/src/main/java/org/dspace/app/rest/submit/factory/impl/LicenseReplacePatchOperation.java
@@ -7,9 +7,7 @@
  */
 package org.dspace.app.rest.submit.factory.impl;
 
-import java.util.List;
-
-import org.dspace.content.Bundle;
+import org.apache.commons.lang.BooleanUtils;
 import org.dspace.content.Item;
 import org.dspace.content.LicenseUtils;
 import org.dspace.content.WorkspaceItem;
@@ -18,18 +16,11 @@ import org.dspace.core.Context;
 import org.dspace.eperson.EPerson;
 import org.dspace.services.model.Request;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.util.Assert;
 
 /**
  * Submission "replace" patch operation
  * 
- * The replace operation allows to replace existent information with new one.
- * Attempt to use the replace operation without a previous accepted license must
- * return an error.
- * 
- * Example: <code>
- * curl -X PATCH http://${dspace.url}/dspace-spring-rest/api/submission/workspaceitems/31599 -H "Content-Type: application/json" -d '[{ "op": "replace", "path": "/sections/license/granted", "value":"true"}]'
- * </code>
+ * {@link LicenseAddPatchOperation}
  * 
  * @author Luigi Andrea Pascarelli (luigiandrea.pascarelli at 4science.it)
  *
@@ -42,20 +33,26 @@ public class LicenseReplacePatchOperation extends ReplacePatchOperation<String> 
 	@Override
 	void replace(Context context, Request currentRequest, WorkspaceItem source, String path, Object value)
 			throws Exception {
-		Item item = source.getItem();
-		List<Bundle> bunds = itemService.getBundles(item, "LICENSE");
-		Assert.notEmpty(bunds);
+		
+		Boolean grant = BooleanUtils.toBooleanObject((String)value);
+		
+		if(grant==null) {
+			throw new IllegalArgumentException("Value is not a valid boolean expression (permitted value: on/off, true/false and yes/no");
+		}
 
+		Item item = source.getItem();
 		EPerson submitter = context.getCurrentUser();
 
 		// remove any existing DSpace license (just in case the user
 		// accepted it previously)
 		itemService.removeDSpaceLicense(context, item);
 
-		String license = LicenseUtils.getLicenseText(context.getCurrentLocale(), source.getCollection(), item,
-				submitter);
-
-		LicenseUtils.grantLicense(context, item, license, null);
+		if(grant) {
+			String license = LicenseUtils.getLicenseText(context.getCurrentLocale(), source.getCollection(), item,
+					submitter);
+	
+			LicenseUtils.grantLicense(context, item, license, null);
+		}
 	}
 
 	@Override
